Feat sc-8646: Move jwt call into it's own service class.
